+++ src/main/kotlin/allegrex/agent/ppsspp/bridge/model/request/PpssppHleBacktraceRequest.kt
@@ -4,7 +4,7 @@ import allegrex.agent.ppsspp.bridge.model.event.PpssppHleBacktraceEvent
 import java.util.UUID
 
 data class PpssppHleBacktraceRequest(
-  val thread: Int,
+  val thread: Long,
   override val ticket: String = UUID.randomUUID().toString(),
 ) : PpssppRequest {
   override val event: String = PpssppHleBacktraceEvent.EVENT_NAME
